pyproject.toml
[project]
name = "specsoloist"
version = "0.2.0"
description = "Spec-as-Source Framework: AI-driven development where specifications are the source of truth."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Toby Lightheart", email = "code@symbolfarm.com" }
]
keywords = ["ai", "llm", "code-generation", "specifications", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "mcp>=0.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/symbolfarm/specsoloist"
Repository = "https://github.com/symbolfarm/specsoloist"
Issues = "https://github.com/symbolfarm/specsoloist/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
sp = "specsoloist.cli:main"
specsoloist-mcp = "specsoloist.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/specsoloist"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
